third_party/WebKit/LayoutTests/http/tests/inspector-protocol/cachestorage/read-cached-response-expected.txt

Page with a service worker
Page reloaded.
Cached requests:
text/html
Type of body: string
application/x-javascript
Type of body: string
Error: Invalid parameters cacheId: string value expected; requestURL: string value expected
Error: Invalid parameters requestURL: string value expected
Error: Invalid cache id. 
Error: Unable to read cached response: not found. 
